Squash /schema GET APIs into single class
